+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+
+- Add optional channelId and publishedAfter/publishedBefore filters to search.
+
 ## [1.1.0] - August 14th 2020
 
 - Initial version of youtube_api_client. (Starting after youtube_api package fork on version 1.0.3).